requirements.txt
redis
3==0.0.0
amqp==5.0.6
asgiref==3.3.1
authy==2.2.6
Babel==2.9.1
billiard==3.6.4.0
celery==5.1.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1.7
django-filter==2.4.0
django-formtools==2.3
django-otp==1.0.6
django-phonenumber-field==1.1.0
django-sendsms==0.2.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.1
djongo==1.3.6
future==0.18.2
gunicorn==20.1.0
idna==2.10
image==1.5.33
kombu==5.1.0
Markdown==3.3.4
phonenumbers==8.12.24
phonenumberslite==8.12.24
Pillow==8.1.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
PyJWT==1.7.1
pymongo==3.12.0
pyotp==2.6.0
pytz==2021.1
qrcode==6.1
requests==2.25.1
Serializer==0.2.1
serializers==0.2.4
simplejson==3.17.2
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.2.4
Unidecode==1.3.2
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
